refactor(confopt): refactor profile to sampler_profile
