+++ scripts/baselines/gdas/gdas_search_and_discretize.py
@@ -5,7 +5,7 @@ import json
 
 import wandb
 
-from confopt.profile import DiscreteProfile, GDASProfile
+from confopt.sampler_profile import DiscreteProfile, GDASProfile
 from confopt.train import Experiment
 from confopt.enums import DatasetType, SearchSpaceType
