+++ scripts/baselines/snas/snas_search.py
@@ -5,7 +5,7 @@ import json
 
 import wandb
 
-from confopt.profile import SNASProfile
+from confopt.sampler_profile import SNASProfile
 from confopt.train import Experiment
 from confopt.enums import DatasetType, SearchSpaceType
