--- src/confopt/profile/__init__.py
@@ -1,19 +1,0 @@
-from .base import BaseProfile
-from .profiles import (
-    DARTSProfile,
-    DiscreteProfile,
-    DRNASProfile,
-    GDASProfile,
-    ReinMaxProfile,
-    SNASProfile,
-)
-
-__all__ = [
-    "BaseProfile",
-    "DARTSProfile",
-    "GDASProfile",
-    "DRNASProfile",
-    "SNASProfile",
-    "DiscreteProfile",
-    "ReinMaxProfile",
-]
